clean up account page stuff
